+++ src/components/forms/checkbox-input.tsx
@@ -1,14 +1,12 @@
 import { Checkbox } from 'antd';
 import { Controller, FieldValues, useFormContext } from 'react-hook-form';
 import { IGenericInputProps, RHFGenericValueType } from '.';
-import { Label } from './label';
 export interface ICheckBoxInputProps<TFieldValues extends FieldValues>
   extends IGenericInputProps<TFieldValues> {
   disabled?: boolean;
 }
 
 export const CheckInput = <TFieldValues extends FieldValues>({
-  tooltip,
   name,
   label,
   disabled,
@@ -20,15 +18,15 @@ export const CheckInput = <TFieldValues extends FieldValues>({
       name={name}
       control={control}
       render={({ field: { value } }) => (
-        <Label label={label} tooltip={tooltip}>
-          <Checkbox
-            onChange={value =>
-              setValue(name, value.target.checked as RHFGenericValueType)
-            }
-            checked={value}
-            disabled={disabled}
-          />
-        </Label>
+        <Checkbox
+          onChange={value =>
+            setValue(name, value.target.checked as RHFGenericValueType)
+          }
+          checked={value}
+          disabled={disabled}
+        >
+          {label}
+        </Checkbox>
       )}
     />
   );